requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bce-python-sdk==0.9.60
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.5
fastapi==0.135.1
filelock==3.25.0
firebase-admin
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ImageIO==2.37.2
imagesize==2.0.0
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==1.26.0
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
prettytable==3.17.0
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.5.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.2
requests==2.32.5
rich==14.3.3
rsa==4.9.1
ruamel.yaml==0.19.1
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tifffile==2026.3.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
pytest==7.4.3
mypy==1.7.1
